# Remove one query from the critical path