With this change, I see a consistent ~15% increase in throughput on my machine (2018 15" Macbook Pro)

`master`:
```

┌─────────┬────────┬────────┬────────┬────────┬───────────┬──────────┬─────────┐
│ Stat    │ 2.5%   │ 50%    │ 97.5%  │ 99%    │ Avg       │ Stdev    │ Max     │
├─────────┼────────┼────────┼────────┼────────┼───────────┼──────────┼─────────┤
│ Latency │ 429 ms │ 469 ms │ 663 ms │ 965 ms │ 490.36 ms │ 86.47 ms │ 1351 ms │
└─────────┴────────┴────────┴────────┴────────┴───────────┴──────────┴─────────┘
┌───────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┐
│ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%   │ Avg     │ Stdev   │ Min     │
├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
│ Req/Sec   │ 13      │ 13      │ 52      │ 57      │ 50.6    │ 8.33    │ 13      │
├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
│ Bytes/Sec │ 1.59 kB │ 1.59 kB │ 6.35 kB │ 6.96 kB │ 6.17 kB │ 1.02 kB │ 1.59 kB │
└───────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┘
```

This branch:
```

┌─────────┬────────┬────────┬────────┬────────┬───────────┬──────────┬─────────┐
│ Stat    │ 2.5%   │ 50%    │ 97.5%  │ 99%    │ Avg       │ Stdev    │ Max     │
├─────────┼────────┼────────┼────────┼────────┼───────────┼──────────┼─────────┤
│ Latency │ 381 ms │ 411 ms │ 573 ms │ 635 ms │ 426.93 ms │ 70.51 ms │ 1343 ms │
└─────────┴────────┴────────┴────────┴────────┴───────────┴──────────┴─────────┘
┌───────────┬────────┬─────────┬─────────┬─────────┬─────────┬───────┬────────┐
│ Stat      │ 1%     │ 2.5%    │ 50%     │ 97.5%   │ Avg     │ Stdev │ Min    │
├───────────┼────────┼─────────┼─────────┼─────────┼─────────┼───────┼────────┤
│ Req/Sec   │ 9      │ 42      │ 61      │ 65      │ 58.35   │ 8.19  │ 9      │
├───────────┼────────┼─────────┼─────────┼─────────┼─────────┼───────┼────────┤
│ Bytes/Sec │ 1.1 kB │ 5.13 kB │ 7.44 kB │ 7.93 kB │ 7.12 kB │ 999 B │ 1.1 kB │
└───────────┴────────┴─────────┴─────────┴─────────┴─────────┴───────┴────────┘
```