@lalexgap 2020-11-19
On CI I'm not seeing much of an improvement with this branch.

From this branch:
```
┌─────────┬────────┬────────┬────────┬────────┬───────────┬──────────┬────────┐
│ Stat    │ 2.5%   │ 50%    │ 97.5%  │ 99%    │ Avg       │ Stdev    │ Max    │
├─────────┼────────┼────────┼────────┼────────┼───────────┼──────────┼────────┤
│ Latency │ 215 ms │ 252 ms │ 374 ms │ 444 ms │ 260.97 ms │ 47.88 ms │ 916 ms │
└─────────┴────────┴────────┴────────┴────────┴───────────┴──────────┴────────┘
┌───────────┬─────────┬────────┬───────┬─────────┬─────────┬─────────┬─────────┐
│ Stat      │ 1%      │ 2.5%   │ 50%   │ 97.5%   │ Avg     │ Stdev   │ Min     │
├───────────┼─────────┼────────┼───────┼─────────┼─────────┼─────────┼─────────┤
│ Req/Sec   │ 26      │ 59     │ 98    │ 110     │ 95.44   │ 13.56   │ 26      │
├───────────┼─────────┼────────┼───────┼─────────┼─────────┼─────────┼─────────┤
│ Bytes/Sec │ 3.17 kB │ 7.2 kB │ 12 kB │ 13.4 kB │ 11.6 kB │ 1.65 kB │ 3.17 kB │
└───────────┴─────────┴────────┴───────┴─────────┴─────────┴─────────┴─────────┘
```

From `master`:
```
┌─────────┬────────┬────────┬────────┬────────┬───────────┬──────────┬────────┐
│ Stat    │ 2.5%   │ 50%    │ 97.5%  │ 99%    │ Avg       │ Stdev    │ Max    │
├─────────┼────────┼────────┼────────┼────────┼───────────┼──────────┼────────┤
│ Latency │ 221 ms │ 250 ms │ 350 ms │ 418 ms │ 259.38 ms │ 41.57 ms │ 819 ms │
└─────────┴────────┴────────┴────────┴────────┴───────────┴──────────┴────────┘
┌───────────┬─────────┬─────────┬───────┬─────────┬─────────┬─────────┬─────────┐
│ Stat      │ 1%      │ 2.5%    │ 50%   │ 97.5%   │ Avg     │ Stdev   │ Min     │
├───────────┼─────────┼─────────┼───────┼─────────┼─────────┼─────────┼─────────┤
│ Req/Sec   │ 34      │ 63      │ 98    │ 109     │ 95.97   │ 12.71   │ 34      │
├───────────┼─────────┼─────────┼───────┼─────────┼─────────┼─────────┼─────────┤
│ Bytes/Sec │ 4.15 kB │ 7.69 kB │ 12 kB │ 13.3 kB │ 11.7 kB │ 1.55 kB │ 4.15 kB │
```