@lalexgap 2020-11-19
On @andrewgordstewart 's machine we're seeing a slight increase in performance so I think it might be worth merging this in.